+++ noodles-tabix/src/async/writer.rs
@@ -1,13 +1,13 @@
 use noodles_bgzf as bgzf;
 use noodles_csi::{
     binning_index::ReferenceSequenceExt,
-    index::reference_sequence::{bin::Chunk, Metadata},
+    index::reference_sequence::{bin::Chunk, Bin, Metadata},
     BinningIndex,
 };
 use tokio::io::{self, AsyncWrite, AsyncWriteExt};
 
 use crate::{
-    index::{header::ReferenceSequenceNames, reference_sequence::Bin, Header, ReferenceSequence},
+    index::{header::ReferenceSequenceNames, Header, ReferenceSequence},
     Index,
 };
 
@@ -289,10 +289,14 @@ async fn write_metadata<W>(writer: &mut W, metadata: &Metadata) -> io::Result<()
 where
     W: AsyncWrite + Unpin,
 {
-    use crate::index::reference_sequence::bin::{METADATA_CHUNK_COUNT, METADATA_ID};
+    use crate::index::DEPTH;
+
+    const METADATA_CHUNK_COUNT: usize = 2;
+
+    let metadata_id = Bin::metadata_id(DEPTH);
 
     let id =
-        u32::try_from(METADATA_ID).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+        u32::try_from(metadata_id).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_u32_le(id).await?;
 
     let n_chunk = u32::try_from(METADATA_CHUNK_COUNT)