+++ noodles-tabix/src/writer.rs
@@ -4,12 +4,12 @@ use byteorder::{LittleEndian, WriteBytesExt};
 use noodles_bgzf as bgzf;
 use noodles_csi::{
     binning_index::ReferenceSequenceExt,
-    index::reference_sequence::{bin::Chunk, Metadata},
+    index::reference_sequence::{bin::Chunk, Bin, Metadata},
     BinningIndex,
 };
 
 use super::{
-    index::{self, header::ReferenceSequenceNames, reference_sequence::Bin, ReferenceSequence},
+    index::{self, header::ReferenceSequenceNames, ReferenceSequence},
     Index, MAGIC_NUMBER,
 };
 
@@ -262,10 +262,14 @@ fn write_metadata<W>(writer: &mut W, metadata: &Metadata) -> io::Result<()>
 where
     W: Write,
 {
-    use crate::index::reference_sequence::bin::{METADATA_CHUNK_COUNT, METADATA_ID};
+    use super::index::DEPTH;
+
+    const METADATA_CHUNK_COUNT: usize = 2;
+
+    let metadata_id = Bin::metadata_id(DEPTH);
 
     let bin_id =
-        u32::try_from(METADATA_ID).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+        u32::try_from(metadata_id).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_u32::<LittleEndian>(bin_id)?;
 
     let n_chunk = u32::try_from(METADATA_CHUNK_COUNT)
@@ -299,7 +303,7 @@ mod tests {
             bgzf::VirtualPosition::from(509268599425),
             bgzf::VirtualPosition::from(509268599570),
         )];
-        let bins = vec![Bin::new(16385, chunks)];
+        let bins = vec![Bin::new(16385, bgzf::VirtualPosition::default(), chunks)];
         let intervals = vec![bgzf::VirtualPosition::from(337)];
         let references = vec![ReferenceSequence::new(bins, intervals, None)];